refactor(streaming): move Pi child process cleanup out of spawn manager
