+++ src/meridian/lib/streaming/spawn_manager.py
@@ -5,8 +5,6 @@ from __future__ import annotations
 import asyncio
 import json
 import logging
-import os
-import signal
 import time
 from collections.abc import Awaitable, Callable
 from contextlib import suppress
@@ -53,6 +51,7 @@ from meridian.lib.streaming.event_observers import (
 )
 from meridian.lib.streaming.heartbeat import heartbeat_loop
 from meridian.lib.streaming.pi_drain import PiDrainCoordinator
+from meridian.lib.streaming.pi_process_cleanup import terminate_pi_tracked_subspawns
 from meridian.lib.streaming.pi_subspawn_tracker import PiSubspawnTracker
 from meridian.lib.streaming.types import InjectResult
 
@@ -421,7 +420,7 @@ class SpawnManager:
             tracker: PiSubspawnTracker,
             reason: str,
         ) -> None:
-            await self._terminate_pi_tracked_subspawns(spawn_id, tracker, reason=reason)
+            await terminate_pi_tracked_subspawns(spawn_id, tracker, reason=reason)
 
         consecutive_write_failures = 0
         max_consecutive_failures = 10
@@ -652,118 +651,6 @@ class SpawnManager:
                         with suppress(asyncio.QueueEmpty):
                             session.subscriber.get_nowait()
                         continue
-
-    async def _terminate_pi_tracked_subspawns(
-        self,
-        spawn_id: SpawnId,
-        tracker: PiSubspawnTracker,
-        *,
-        reason: str,
-    ) -> None:
-        pgids = tracker.active_tracked_pgid_candidates()
-        if not pgids:
-            logger.warning(
-                "Pi spawn %s ended with tracked children but no pid/pgid metadata for cleanup",
-                spawn_id,
-            )
-            return
-
-        for pgid in pgids:
-            if os.name == "nt":
-                await self._terminate_process_tree_fallback(
-                    spawn_id=spawn_id,
-                    process_id=pgid,
-                    reason=reason,
-                )
-            else:
-                await self._terminate_posix_process_group(
-                    spawn_id=spawn_id,
-                    process_group_id=pgid,
-                    reason=reason,
-                )
-
-    async def _terminate_process_tree_fallback(
-        self,
-        *,
-        spawn_id: SpawnId,
-        process_id: int,
-        reason: str,
-    ) -> None:
-        if process_id <= 0:
-            return
-
-        try:
-            from meridian.lib.platform.process_scope.fallback import terminate_tree_sync
-
-            await asyncio.to_thread(
-                terminate_tree_sync,
-                pid=process_id,
-                grace_secs=5.0,
-                reason=reason,
-                scope_id=f"pi-subspawn:{spawn_id}",
-                degraded_fallback=True,
-            )
-        except Exception:
-            logger.warning(
-                "Failed fallback cleanup for Pi child process %d (spawn %s, reason=%s)",
-                process_id,
-                spawn_id,
-                reason,
-                exc_info=True,
-            )
-
-    async def _terminate_posix_process_group(
-        self,
-        *,
-        spawn_id: SpawnId,
-        process_group_id: int,
-        reason: str,
-    ) -> None:
-        if os.name == "nt" or process_group_id <= 0:
-            return
-
-        try:
-            os.killpg(process_group_id, signal.SIGTERM)
-        except ProcessLookupError:
-            return
-        except (PermissionError, OSError):
-            logger.warning(
-                "Failed SIGTERM cleanup for Pi child process group %d (spawn %s, reason=%s)",
-                process_group_id,
-                spawn_id,
-                reason,
-                exc_info=True,
-            )
-            return
-
-        await asyncio.sleep(0.25)
-
-        try:
-            os.killpg(process_group_id, 0)
-        except ProcessLookupError:
-            return
-        except (PermissionError, OSError):
-            logger.warning(
-                "Failed liveness check for Pi child process group %d (spawn %s, reason=%s)",
-                process_group_id,
-                spawn_id,
-                reason,
-                exc_info=True,
-            )
-            return
-
-        try:
-            os.killpg(process_group_id, signal.SIGKILL)
-        except ProcessLookupError:
-            return
-        except (PermissionError, OSError):
-            logger.warning(
-                "Failed SIGKILL cleanup for Pi child process group %d (spawn %s, reason=%s)",
-                process_group_id,
-                spawn_id,
-                reason,
-                exc_info=True,
-            )
 
     def subscribe(self, spawn_id: SpawnId) -> asyncio.Queue[HarnessEvent | None] | None:
         """Attach one subscriber queue to the spawn, or return None if unavailable."""
